backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-docx==0.8.11
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2